feat: Display active miners in header
This commit introduces a new component, `ActiveMiners`, to display the number of active miners. It also updates the Mainnet launch date in the FAQ section and refactors the MinersCTA component to use the new ActiveMiners component.
This change aims to improve the user experience by providing real-time information about the network activity.