fix: Update favicon path in layout metadata
The favicon path in the layout metadata was incorrectly pointing to `https://tari.com//favicon.png?v=1` (notice the double slash). This commit corrects the path to `https://tari.com/favicon.png?v=1` to ensure the favicon is correctly loaded. Also, a favicon.ico file was added to the project.