feat: Add GTM event tracking to download buttons
This commit adds Google Tag Manager (GTM) event tracking to the download buttons for MacOS, Windows, and Linux. When a user clicks on a download button, a `download_button_clicked` event is sent to GTM, along with the platform (macos, windows, or linux) that was clicked. This will allow us to track which platforms are being downloaded most frequently.