refactor: Move swap components to shared directory
Moves swap related components from tari-dot-com to a shared "Swap" directory. This is done to prepare for reuse of the swap functionality in other parts of the application, reducing duplication and improving maintainability.

Additionally, this commit introduces a new helper function `formatNumberWithCommas` to format number inputs with commas, enhancing readability for users. It also include the `cleanFormattedNumber` to allow handle only number value for calculations.