feat: Add wXTM to wallet function
Adds a function `addXtmToWallet` to prompt users to add wXTM (wrapped XTM) to their wallet using the `wallet_watchAsset` method.  The function is called within the Swap component to improve user experience, especially for users interacting with wXTM for the first time. Wrapped XTM address and details are hardcoded to avoid the need for a token object.